test(p2p): move libp2p protocol tests into the tests tree
