+++ crates/p2p/src/two_stream/handler/manage.rs
@@ -238,20 +238,3 @@ impl TwoStreamHandler {
         Ok(TwoStreamEvent::ManageQueryReply { peer_id, reply })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    #[tokio::test]
-    async fn manage_request_decodes() {
-        use crate::message::{ManageMutateOp, ManageRequest};
-        let req = ManageRequest::new(
-            ManageMutateOp::CollectionAdd {
-                collection_ids: vec!["c1".into()],
-            },
-            b"t".to_vec(),
-        );
-        let back: ManageRequest =
-            defra_core::cbor::from_slice(&defra_core::cbor::to_vec(&req).unwrap()).unwrap();
-        assert!(matches!(back.op, ManageMutateOp::CollectionAdd { .. }));
-    }
-}
