+++ crates/p2p/tests/libp2p_protocol/main.rs
@@ -1,0 +1,10 @@
+#![cfg(feature = "libp2p-transport")]
+
+mod pubsub_correlator;
+mod pubsub_envelope;
+mod pubsub_id;
+mod pubsub_topic;
+mod pubsub_topic_handler;
+mod two_stream;
+mod two_stream_event;
+mod two_stream_manage;
