+++ crates/p2p/tests/libp2p_protocol/two_stream_event.rs
@@ -1,0 +1,15 @@
+use p2p::TwoStreamEvent;
+
+#[test]
+fn manage_variants_exist() {
+    fn _a(e: TwoStreamEvent) -> bool {
+        matches!(
+            e,
+            TwoStreamEvent::ManageRequest { .. }
+                | TwoStreamEvent::ManageReply { .. }
+                | TwoStreamEvent::ManageQueryRequest { .. }
+                | TwoStreamEvent::ManageQueryReply { .. }
+        )
+    }
+    let _ = _a;
+}
